// functions/src/types/Activity.ts
export interface ActivityData {
  date:                  string, // YYYY-MM-DD
  weekday:               string, // For visualization in frontend (set on POST/PUT in route)
  startTime:             string, // HH24:MM
  endTime:               string, // HH24:MM
  color:                 string, // For visualization in frontend
  period:                string,
  description:           string, // Free text field

  type:                  string, // (Reference) If generated by an Activity type
  typeName:              string, // (Reference) If generated by an Activity type

  // Availability
  availableToUsers:      boolean,       // If this should be available to users (to book themselves)
  typelessSlots:         boolean,       // If slot is without type (i.e. single slot, not an array)
  assistantSlots:        Array<string>, // Array of slots in increasing assistant level, or single slot (for typeless)

  // Booking data
  assistantAllocations:  Array<string>, // Array of allocated in increasing assistant level, or single slot (for typeless)
  acceptedActivityBookings: Array<string>, // List of accepted activity bookings (id)
}

export interface Activity extends ActivityData {
  id: string
}

export const EMPTY_ACTIVITY = {
  id:                       "",
  date:                     "",
  weekday:                  "",
  startTime:                "",
  endTime:                  "",
  color:                    "",
  period:                   "",
  description:              "",
  type:                     "",
  typeName:                 "",
  availableToUsers:         false,
  typelessSlots:            false,
  assistantSlots:           [],
  assistantAllocations:     [],
  acceptedActivityBookings: [],
};